[#107] refactor(profile): 다른 도메인에서 Profile 사용하도록 변경
- Post 도메인: Gender, AgeRange import 경로 변경 (user.enums → profile.domain.enums)
- PostRepositoryImpl: InvalidGenderException import 경로 변경
- CommentResponse: User 대신 Profile에서 nickname, profileImage 조회
- UserInfo: User.getProfile()을 통해 Profile 정보 조회하도록 변경
- PhoneVerificationService: UserRepository 대신 ProfileRepository 사용
