+++ src/main/java/com/dduru/gildongmu/post/domain/Post.java
@@ -4,10 +4,15 @@ import com.dduru.gildongmu.common.entity.BaseTimeEntity;
 import com.dduru.gildongmu.destination.domain.Destination;
 import com.dduru.gildongmu.participation.domain.Participation;
 import com.dduru.gildongmu.post.enums.PostStatus;
-import com.dduru.gildongmu.post.exception.*;
+import com.dduru.gildongmu.post.exception.InvalidPostStatusException;
+import com.dduru.gildongmu.post.exception.InvalidRecruitCapacityException;
+import com.dduru.gildongmu.post.exception.RecruitCountBelowZeroException;
+import com.dduru.gildongmu.post.exception.RecruitCountExceedCapacityException;
+import com.dduru.gildongmu.post.exception.TravelAlreadyEndedException;
+import com.dduru.gildongmu.post.exception.TravelAlreadyStartedException;
+import com.dduru.gildongmu.profile.domain.enums.AgeRange;
+import com.dduru.gildongmu.profile.domain.enums.Gender;
 import com.dduru.gildongmu.user.domain.User;
-import com.dduru.gildongmu.user.enums.AgeRange;
-import com.dduru.gildongmu.user.enums.Gender;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
