+++ src/main/java/com/dduru/gildongmu/post/repository/PostRepositoryImpl.java
@@ -1,11 +1,11 @@
 package com.dduru.gildongmu.post.repository;
 
-import com.dduru.gildongmu.user.enums.AgeRange;
-import com.dduru.gildongmu.user.enums.Gender;
 import com.dduru.gildongmu.post.domain.Post;
 import com.dduru.gildongmu.post.dto.PostListRequest;
 import com.dduru.gildongmu.post.enums.PostStatus;
-import com.dduru.gildongmu.user.exception.InvalidGenderException;
+import com.dduru.gildongmu.profile.domain.enums.AgeRange;
+import com.dduru.gildongmu.profile.domain.enums.Gender;
+import com.dduru.gildongmu.profile.exception.InvalidGenderException;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
