+++ src/main/java/com/dduru/gildongmu/user/dto/UserInfo.java
@@ -1,5 +1,6 @@
 package com.dduru.gildongmu.user.dto;
 
+import com.dduru.gildongmu.profile.domain.Profile;
 import com.dduru.gildongmu.user.domain.User;
 
 public record UserInfo(
@@ -11,13 +12,14 @@ public record UserInfo(
         String nickname
 ) {
     public static UserInfo from(User user) {
+        Profile profile = user.getProfile();
         return new UserInfo(
                 user.getId(),
                 user.getName(),
-                user.getProfileImage(),
-                user.getGender() != null ? user.getGender().name() : null,
-                user.getBirthday() != null ? user.getBirthday().toString() : null,
-                user.getNickname()
+                profile != null ? profile.getProfileImage() : null,
+                profile != null && profile.getGender() != null ? profile.getGender().name() : null,
+                profile != null && profile.getBirthday() != null ? profile.getBirthday().toString() : null,
+                profile != null ? profile.getNickname() : null
         );
     }
 }
